requirements.txt
flask==2.2.3
requests==2.28.2
twilio==7.17.0
gunicorn==20.1.0
uvicorn==0.21.1
openai==0.27.2
google-cloud-storage
google-cloud-storage-transfer
deepgram-sdk==2.3.0



